+++ lib/Clarity/parse.mli
@@ -3,3 +3,5 @@
 val parse_program : string -> program
 
 val parse_definition : Sexplib.Sexp.t -> definition
+
+val parse_type : Sexplib.Sexp.t -> type'