+++ lib/Clarity/print.ml
@@ -52,8 +52,10 @@ and print_literal ppf = function
   | StringLiteral s -> fprintf ppf "\"%s\"" s  (* TODO: escaping *)
 
 and type_to_string = function
-  | Principal -> "principal"
   | Bool -> "bool"
   | Int -> "int"
   | Uint -> "uint"
-  | String -> "string"
+  | Principal -> "principal"
+  | Optional _ -> "optional"  (* TODO *)
+  | String _ -> "string"  (* TODO *)
+  | List _ -> "list"  (* TODO *)