+++ lib/SWIR/print.mli
@@ -3,3 +3,7 @@
 val print_program : Format.formatter -> program -> unit
 
 val print_definition : Format.formatter -> definition -> unit
+
+val print_type : Format.formatter -> type' -> unit
+
+val type_to_string : type' -> string