+++ CHANGES.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for `define-constant`
 - Support for `define-map`
+- Support for `optional` types
+- Support for `string-ascii` and `string-utf8` types
+- Support for `list` types
 
 ## 0.1.0 - 2020-09-18